backend/requirements.txt
fastapi 
uvicorn[standard] 
python-multipart 
pypdf 
python-docx 
requests 
openai          
python-dotenv   